refactor(backend): :recycle: split authentication and CRUD functions in different files
Refs #9
