+++ backend/src/main/java/com/tfg/cultura/api/users/service/UserService.java
@@ -1,148 +1,32 @@
 package com.tfg.cultura.api.users.service;
 
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.DisabledException;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-
-import com.tfg.cultura.api.users.repository.UserRepository;
-
-import com.tfg.cultura.api.users.model.User;
-import com.tfg.cultura.api.users.model.dto.*;
-import com.tfg.cultura.api.core.exception.UnathenticatedException;
-import com.tfg.cultura.api.users.exception.*;
-import com.tfg.cultura.api.users.jwt.CustomUserDetails;
-import com.tfg.cultura.api.users.jwt.CustomUserDetailsService;
-import com.tfg.cultura.api.users.jwt.JwtService;
-
 import java.util.Optional;
 
-import com.tfg.cultura.api.core.exception.FileUploadException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
+
+import com.tfg.cultura.api.users.exception.UserNotFoundException;
+import com.tfg.cultura.api.users.model.User;
+import com.tfg.cultura.api.users.model.dto.UserResponse;
+import com.tfg.cultura.api.users.repository.UserRepository;
 
 @Service
 public class UserService {
-
     private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-    private final JwtService jwtService;
-    private final CustomUserDetailsService userDetailsService;
-    private final UserFileService userFileService;
 
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder, JwtService jwtService,
-            CustomUserDetailsService userDetailsService, UserFileService userFileService) {
+    public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
-        this.passwordEncoder = passwordEncoder;
-        this.jwtService = jwtService;
-        this.userDetailsService = userDetailsService;
-        this.userFileService = userFileService;
     }
 
     private static final Logger logger = LoggerFactory.getLogger("usersLogger");
-
-    public UserResponse register(UserRegisterRequest request, MultipartFile avatar, MultipartFile paymentReceipt) throws UserAlreadyExistsException, FileUploadException {
-        userFileService.validateAvatar(avatar);
-        userFileService.validatePaymentReceipt(paymentReceipt);
-
-        if (userRepository.existsByUsername(request.getUsername())) {
-            logger.warn("Error al registrar el usuario {}: El nombre de usuario ya existe", request.getUsername());
-            throw new UserAlreadyExistsException("El nombre de usuario ya existe");
-        }
-
-        if (userRepository.existsByDni(request.getDni())) {
-            logger.warn("Error al registrar el usuario {}: El DNI {} ya existe", request.getUsername(),
-                    request.getDni());
-            throw new UserAlreadyExistsException("Ya existe un usuario con el mismo DNI");
-        }
-
-        User user = User.builder()
-            .username(request.getUsername())
-            .password(passwordEncoder.encode(request.getPassword()))
-            .name(request.getName())
-            .surname(request.getSurname())
-            .dni(request.getDni())
-            .phone(request.getPhone())
-            .email(request.getEmail())
-            .avatar(UserFileService.AVATAR_PLACEHOLDER)
-            .build();
-
-        User savedUser = userRepository.save(user);
-        String userId = savedUser.getId();
-        logger.info("Usuario registrado correctamente: {}", savedUser.getUsername());
-
-        if (avatar != null && !avatar.isEmpty()) {
-            logger.info("Se va a intentar subir el avatar: {}", avatar.getOriginalFilename());
-            String avatarUrl = userFileService.uploadAvatar(userId, avatar);
-            user.setAvatar(avatarUrl);
-        }
-
-        logger.info("Se va a intentar subir el PDF de la carta de pago: {}",
-                paymentReceipt.getOriginalFilename());
-        String paymentReceiptUrl = userFileService.uploadPaymentReceiptPdf(userId, paymentReceipt);
-        user.setPaymentReceipt(paymentReceiptUrl);
-
-        return new UserResponse(savedUser);
-    }
-
-    public String login(UserLoginRequest request)
-            throws UserNotFoundException, DisabledException, BadCredentialsException {
-        Optional<User> user = userRepository.findByUsername(request.getUsername());
-        if (user.isEmpty()) {
-            logger.warn("Error al iniciar sesión: El usuario {} no existe", request.getUsername());
-            throw new UserNotFoundException(
-                    "El usuario con username " + request.getUsername() + " no existe");
-        }
-        User foundUser = user.get();
-        if (!foundUser.isActive()) {
-            logger.warn("Error al iniciar sesión: El usuario {} está desactivado", request.getUsername());
-            throw new DisabledException("El usuario está desactivado");
-        }
-
-        if (!passwordEncoder.matches(request.getPassword(), foundUser.getPassword())) {
-            logger.warn("Error al iniciar sesión: El usuario {} introdujo una contraseña incorrecta",
-                    request.getUsername());
-            throw new BadCredentialsException("Credenciales inválidas");
-        }
-
-        CustomUserDetails userDetails = (CustomUserDetails) userDetailsService
-                .loadUserByUsername(request.getUsername());
-
-        return jwtService.generateToken(
-                userDetails.getUsername(),
-                userDetails.getRole(),
-                userDetails.getId());
-    }
 
     public UserResponse getUserById(String id) throws UserNotFoundException {
         User user = findUserById(id);
         return new UserResponse(user);
     }
 
-    public UserResponse activateUser(String id) throws UserNotFoundException {
-
-        User user = findUserById(id);
-        CustomUserDetails currentUser = getCurrentUserDetails();
-
-        if (user.getId().equals(currentUser.getId())) {
-            throw new SelfActivationNotAllowedException(
-                    String.format("El usuario %s con id %s ha intentado activar su propio usuario", user.getUsername(),
-                            user.getId()));
-        }
-
-        if (!user.isActive()) {
-            user.setActive(true);
-            user = userRepository.save(user);
-        }
-
-        logger.info("Se ha aprobado el registro del usuario {} con id {}", user.getUsername(), user.getId());
-        return new UserResponse(user);
-    }
-
-    private User findUserById(String id) throws UserNotFoundException {
+    User findUserById(String id) throws UserNotFoundException {
         Optional<User> user = userRepository.findById(id);
 
         if (user.isEmpty()) {
@@ -152,19 +36,5 @@ public class UserService {
 
         return user.get();
     }
-
-    private CustomUserDetails getCurrentUserDetails() {
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        if (auth == null || !auth.isAuthenticated()) {
-            throw new UnathenticatedException("No se ha podido obtener la autenticación del usuario");
-        }
-
-        CustomUserDetails currentUser = (CustomUserDetails) auth.getPrincipal();
-        if (currentUser == null) {
-            throw new UnathenticatedException("No se ha podido obtener la información del usuario");
-        }
-
-        return currentUser;
-    }
-
+    
 }
