+++ backend/src/test/java/com/tfg/cultura/api/users/controller/UserControllerTest.java
@@ -1,32 +1,22 @@
 package com.tfg.cultura.api.users.controller;
 
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.when;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.DisabledException;
 import org.springframework.test.web.servlet.MockMvc;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-import com.tfg.cultura.api.core.exception.UnathenticatedException;
-import com.tfg.cultura.api.users.exception.SelfActivationNotAllowedException;
-import com.tfg.cultura.api.users.exception.UserAlreadyExistsException;
 import com.tfg.cultura.api.users.exception.UserNotFoundException;
 import com.tfg.cultura.api.users.exception.UsersExceptionHandler;
 import com.tfg.cultura.api.users.factory.UserFactory;
-import com.tfg.cultura.api.users.model.dto.UserLoginRequest;
-import com.tfg.cultura.api.users.model.dto.UserRegisterRequest;
 import com.tfg.cultura.api.users.model.dto.UserResponse;
 import com.tfg.cultura.api.users.service.UserService;
 import com.tfg.cultura.api.utils.BaseControllerTest;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 class UserControllerTest extends BaseControllerTest {
 
@@ -36,13 +26,8 @@ class UserControllerTest extends BaseControllerTest {
     private UserService userService;
 
     private static final String BASE_URL = "/api/users";
-    private static final String REGISTER_URL = BASE_URL + "/register";
-    private static final String LOGIN_URL = BASE_URL + "/login";
     private static final String USER_URL = BASE_URL + "/{id}";
-    private static final String ACTIVATE_URL = USER_URL + "/activate";
 
-    private UserRegisterRequest registerRequest;
-    private UserLoginRequest loginRequest;
     private UserResponse userResponse;
 
     @BeforeEach
@@ -55,163 +40,7 @@ class UserControllerTest extends BaseControllerTest {
     }
 
     private void initTestData() {
-        registerRequest = UserFactory.validUserRegisterRequest();
         userResponse = UserFactory.validUserResponse();
-        loginRequest = UserFactory.loginRequest();
-    }
-
-    @Test
-    void register_success() throws Exception {
-        when(userService.register(any(),any(),any())).thenReturn(userResponse);
-
-        mockMvc.perform(multipart(REGISTER_URL)
-                .file(userPart(registerRequest))
-                .file(pdfPart()))
-                .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.username").value(registerRequest.getUsername()));
-    }
-
-    @Test
-    void register_fail_user_already_exists() throws Exception {
-        UserAlreadyExistsException ex = new UserAlreadyExistsException("El nombre de usuario ya existe");
-
-        when(userService.register(any(),any(),any())).thenThrow(ex);
-
-        mockMvc.perform(multipart(REGISTER_URL)
-                .file(userPart(registerRequest))
-                .file(pdfPart()))
-                .andExpect(status().isConflict())
-                .andExpect(jsonPath("$.message").exists());
-    }
-
-    @Test
-    void register_fail_invalid_data() throws Exception {
-        registerRequest.setEmail("invalid-email");
-
-        mockMvc.perform(multipart(REGISTER_URL)
-                .file(userPart(registerRequest))
-                .file(pdfPart()))
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("$.message").exists());
-    }
-
-    // ====== LOGIN ========
-
-    @Test
-    void login_success() throws Exception {
-        String token = "token123";
-        when(userService.login(any())).thenReturn(token);
-
-        mockMvc.perform(post(LOGIN_URL)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJson(loginRequest)))
-                .andExpect(status().isOk())
-                .andExpect(content().string(token));
-
-    }
-
-    @Test
-    void login_fail_invalid_credentials() throws Exception {
-        BadCredentialsException ex = new BadCredentialsException("Credenciales inválidas");
-        when(userService.login(any())).thenThrow(ex);
-
-        mockMvc.perform(post(LOGIN_URL)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJson(loginRequest)))
-                .andExpect(status().isUnauthorized())
-                .andExpect(jsonPath("$.message").value(ex.getMessage()));
-
-    }
-
-    @Test
-    void login_fail_user_not_found() throws Exception {
-        String username = loginRequest.getUsername();
-        String message = "El usuario con username " + username + " no existe";
-        UserNotFoundException ex = new UserNotFoundException(message);
-        when(userService.login(any())).thenThrow(ex);
-
-        mockMvc.perform(post(LOGIN_URL)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJson(loginRequest)))
-                .andExpect(status().isNotFound())
-                .andExpect(jsonPath("$.message").value(message));
-    }
-
-    @Test
-    void login_fail_user_disabled() throws Exception {
-        String message = "El usuario está desactivado";
-        DisabledException ex = new DisabledException(message);
-        when(userService.login(any())).thenThrow(ex);
-
-        mockMvc.perform(post(LOGIN_URL)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJson(loginRequest)))
-                .andExpect(status().isForbidden())
-                .andExpect(jsonPath("$.message").value(message));
-    }
-
-    // Helpers
-
-    private MockMultipartFile userPart(Object obj) throws Exception {
-        return new MockMultipartFile(
-                "user",
-                "",
-                "application/json",
-                toJson(obj).getBytes());
-    }
-
-    private MockMultipartFile pdfPart() {
-        return new MockMultipartFile(
-                "paymentReceipt",
-                "test.pdf",
-                "application/pdf",
-                "dummy pdf content".getBytes());
-    }
-
-    // ================ ACTIVATE USER ================
-
-    @Test
-    void activate_user_sucess() throws Exception {
-        when(userService.activateUser(any())).thenReturn(userResponse);
-
-        mockMvc.perform(put(ACTIVATE_URL, "123"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.username").value(userResponse.getUsername()));
-    }
-
-    @Test
-    void activate_user_fail_unexisting_user() throws Exception {
-        String message = "El usuario con id 123 no existe";
-        UserNotFoundException ex = new UserNotFoundException(message);
-        when(userService.activateUser(any())).thenThrow(ex);
-
-        mockMvc.perform(put(ACTIVATE_URL, "123"))
-                .andExpect(status().isNotFound())
-                .andExpect(jsonPath("$.message").value(message));
-    }
-
-    @Test
-    void activate_user_fail_self_activation() throws Exception {
-        String userId = "123";
-        String message = String.format("El usuario %s con id %s ha intentado activar su propio usuario",
-                userResponse.getUsername(), userId);
-        SelfActivationNotAllowedException ex = new SelfActivationNotAllowedException(message);
-        when(userService.activateUser(userId)).thenThrow(ex);
-
-        mockMvc.perform(put(ACTIVATE_URL, userId))
-                .andExpect(status().isForbidden())
-                .andExpect(jsonPath("$.message").value(message));
-    }
-
-    @Test
-    void activate_user_fail_unathenticated() throws Exception {
-        String message = "No se ha podido obtener la autenticación del usuario";
-        UnathenticatedException ex = new UnathenticatedException(message);
-        when(userService.activateUser(any())).thenThrow(ex);
-
-        mockMvc.perform(put(ACTIVATE_URL, "123"))
-                .andExpect(status().isUnauthorized())
-                .andExpect(jsonPath("$.message").value(message));
     }
 
     // ================ GET USER ================
@@ -235,5 +64,5 @@ class UserControllerTest extends BaseControllerTest {
                 .andExpect(status().isNotFound())
                 .andExpect(jsonPath("$.message").value(message));
     }
-
+    
 }
